refactor(reports): Reorganized modules
