+++ reports/pyproject.toml
@@ -9,10 +9,11 @@ readme = "README.md"
 python = "^3.10"
 fastapi = "^0.88.0"
 uvicorn = {extras = ["standard"], version = "^0.20.0"}
-db-handler = {path = "../libs/db-handler"}
 python-dotenv = "^0.21.0"
 starlette-prometheus = "^0.9.0"
 email-validator = "^1.3.0"
+db-handler = {path = "../libs/db-handler"}
+query = {path = "../libs/query"}
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.2.0"
