+++ reports/tests/test_get_single_report.py
@@ -2,7 +2,7 @@ from unittest import mock
 
 from pymongo.errors import ServerSelectionTimeoutError
 
-from reports.models import Report
+from reports.database import Report
 from . import utils
 
 oid = utils.random_oid()
@@ -10,7 +10,7 @@ report = utils.report_factory(_id=oid)
 endpoint = f"/{oid}"
 
 
-@mock.patch('reports.database.get_connection')
+@mock.patch('reports.database._interactions.get_connection')
 def test_read_report_success(mock_connection):
     find_one = mock.AsyncMock()
     find_one.return_value = report
@@ -21,7 +21,7 @@ def test_read_report_success(mock_connection):
     find_one.assert_awaited_once_with(Report, {"_id": oid})
 
 
-@mock.patch('reports.database.get_connection')
+@mock.patch('reports.database._interactions.get_connection')
 def test_read_report_fails_if_missing_document(mock_connection):
     find_one = mock.AsyncMock()
     find_one.return_value = None
@@ -31,7 +31,7 @@ def test_read_report_fails_if_missing_document(mock_connection):
     assert response.status_code == 404
 
 
-@mock.patch('reports.database.get_connection')
+@mock.patch('reports.database._interactions.get_connection')
 def test_read_report_fails_if_database_is_down(mock_connection):
     find_one = mock.AsyncMock()
     find_one.side_effect = ServerSelectionTimeoutError()
