+++ reports/tests/test_put_report.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 from pymongo.errors import DuplicateKeyError, ServerSelectionTimeoutError
 
-from reports.models import Report
+from reports.database import Report
 from . import utils
 
 oid = utils.random_oid()
@@ -11,7 +11,7 @@ report = utils.report_factory()
 endpoint = f"/{oid}"
 
 
-@mock.patch('reports.database.get_connection')
+@mock.patch('reports.database._interactions.get_connection')
 def test_put_ignores_fields_not_defined_in_schema(mock_connection):
     update_one = mock.AsyncMock()
     update_one.return_value = report
@@ -30,7 +30,7 @@ def test_put_fails_if_missing_fields_defined_in_schema():
     assert response.status_code == 422
 
 
-@mock.patch('reports.database.get_connection')
+@mock.patch('reports.database._interactions.get_connection')
 def test_put_fails_if_missing_document(mock_connection):
     update_one = mock.AsyncMock()
     update_one.return_value = None
@@ -40,7 +40,7 @@ def test_put_fails_if_missing_document(mock_connection):
     assert response.status_code == 404
 
 
-@mock.patch('reports.database.get_connection')
+@mock.patch('reports.database._interactions.get_connection')
 def test_put_report_duplicate_fails(mock_connection):
     update_one = mock.AsyncMock()
     update_one.side_effect = DuplicateKeyError(error="")
@@ -50,7 +50,7 @@ def test_put_report_duplicate_fails(mock_connection):
     assert response.status_code == 400
 
 
-@mock.patch('reports.database.get_connection')
+@mock.patch('reports.database._interactions.get_connection')
 def test_put_report_fails_if_database_is_down(mock_connection):
     update_one = mock.AsyncMock()
     update_one.side_effect = ServerSelectionTimeoutError()
